refactor posts endpoints by public api

+++ src/modules/blogs/api/blogs.blogger.controller.ts
@@ -4,6 +4,7 @@ import {
   Delete,
   Get,
   HttpCode,
+  NotFoundException,
   Param,
   Post,
   Put,
@@ -53,7 +54,7 @@ export class BlogsBloggerController {
   }
 
   @Post('/:blogId/posts')
-  @UseGuards(CanUserWorkWithBlog) //!можно ли через новый конструктор передавать аргумент id или как я сделал в самом гварде на 20 строке, можно ли вообще так делать / blogId  //? спросить можно ли с useCase выкидывать Exception ?
+  @UseGuards(CanUserWorkWithBlog)
   async createPostByBlog(
     @ReqUser() user: DbType<User>,
     @Param('blogId') blogId: string,
@@ -100,6 +101,10 @@ export class BlogsBloggerController {
   @UseGuards(CanUserWorkWithBlog)
   @HttpCode(204)
   async deletePostByBlog(@Param('postId') postId: string) {
+    const post = await this.postsQueryRepo.getPostById(postId, null);
+
+    if (!post) throw new NotFoundException();
+
     return await this.blogsService.deletePostByBlog(postId);
   }
 }
